% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{assig}
\alias{assig}
\title{Assign simulation setting indices}
\usage{
assig(n_args)
}
\arguments{
\item{n_args}{An integer vector giving the number of candidate values for each
simulation setting factor.}
}
\value{
A matrix whose columns correspond to simulation jobs and whose rows
correspond to simulation setting factors. Each column gives one combination
of setting indices.
}
\description{
Generates all combinations of indices for a collection of simulation setting
factors.
}
\details{
This function is used to map a job number to a specific combination of
simulation settings. For example, if \code{n_args = c(2, 3)}, then the function
returns all index combinations from \verb{\{1, 2\} x \{1, 2, 3\}}.
}
\examples{
assig(c(2, 3))

}
