% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{build_ar_sigma}
\alias{build_ar_sigma}
\title{Build an autoregressive covariance matrix}
\usage{
build_ar_sigma(rho, p)
}
\arguments{
\item{rho}{A numeric value specifying the autoregressive correlation
parameter.}

\item{p}{An integer specifying the dimension of the covariance matrix.}
}
\value{
A \code{p} by \code{p} covariance matrix with entries
\eqn{\Sigma_{ij} = \rho^{|i-j|}}.
}
\description{
Constructs an autoregressive covariance matrix for simulation studies.
}
\details{
This function is used to generate covariance matrices under the AR structure
in the simulation studies.
}
\examples{
Sigma <- build_ar_sigma(rho = 0.5, p = 10)

}
