% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{cgztest}
\alias{cgztest}
\title{Zhong and Chen global test}
\usage{
cgztest(X, y, maxabsy = exp(10))
}
\arguments{
\item{X}{A numeric covariate matrix with rows corresponding to observations
and columns corresponding to covariates.}

\item{y}{A numeric response vector.}

\item{maxabsy}{A numeric threshold used to truncate extremely large absolute
response values. Default is \code{exp(10)}.}
}
\value{
A list containing:
\item{Tn}{The standardized test statistic.}
\item{pvals}{The one-sided p-value of the test.}
}
\description{
Implements the high-dimensional global test of Zhong and Chen (2011) for
uncensored data.
}
\details{
This function calls the C routine \verb{_ZC_Test} to compute the test statistic and
variance components. If the response contains extremely large values, it is
truncated at \code{maxabsy} to avoid numerical instability.
}
\examples{
X <- matrix(rnorm(100 * 20), 100, 20)
y <- rnorm(100)
res <- zctest(X, y)

}
\references{
Zhong, P.-S. and Chen, S. X. (2011). Tests for high-dimensional regression
coefficients with factorial designs. \emph{Journal of the American
Statistical Association}, 106(493), 260--274.
}
