% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{pvalrp}
\alias{pvalrp}
\title{Random projection global test}
\usage{
pvalrp(x, y, rho = 0.4, Pk = NULL)
}
\arguments{
\item{x}{A numeric covariate matrix with rows corresponding to observations
and columns corresponding to covariates.}

\item{y}{A numeric response vector.}

\item{rho}{A numeric value specifying the projection ratio. Default is \code{0.4}.}

\item{Pk}{An optional projection matrix. If \code{NULL}, a random Gaussian
projection matrix is generated internally when projection is needed.}
}
\value{
A list containing:
\item{Tn}{The standardized test statistic.}
\item{pvals}{The p-value of the test.}
\item{rho}{The projection ratio used in the test.}
}
\description{
Implements a random projection-based global test for high-dimensional linear
regression models.
}
\details{
This function implements a random projection-based global test for testing
whether the regression coefficients of \code{x} are zero. If the number of
covariates is smaller than \code{rho * n}, the test is computed without random
projection. Otherwise, the covariates are projected to a lower-dimensional
space using either the user-provided projection matrix \code{Pk} or a randomly
generated Gaussian projection matrix.

The core test statistic is computed by C routines registered in the package
and then standardized using the normal approximation.
}
\examples{
x <- matrix(rnorm(100 * 50), 100, 50)
y <- rnorm(100)
res <- pvalrp(x, y, rho = 0.4)
res$pvals

}
