% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{revcumprod}
\alias{revcumprod}
\title{Compute reverse cumulative products}
\usage{
revcumprod(vec)
}
\arguments{
\item{vec}{A numeric vector.}
}
\value{
A numeric vector of the same length as \code{vec}. The first element is
the last element of \code{vec}, and each subsequent element is the cumulative
product of elements from the end of \code{vec} moving backward.
}
\description{
Computes cumulative products of a numeric vector in reverse order.
}
\details{
For a vector \code{vec = c(v1, v2, ..., vn)}, this function returns
\verb{c(vn, vn * v_\{n-1\}, ..., vn * v_\{n-1\} * ... * v1)}.
}
\examples{
revcumprod(c(2, 3, 4))

}
