% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{set_p}
\alias{set_p}
\title{Set the covariate dimension}
\usage{
set_p(n)
}
\arguments{
\item{n}{An integer specifying the sample size.}
}
\value{
An integer giving the covariate dimension.
}
\description{
Computes the covariate dimension used in the simulation studies as a function
of the sample size.
}
\details{
The dimension is calculated as
\deqn{p = \mathrm{round}\{\exp((2n/3)^{0.4}) + 100\}.}
}
\examples{
set_p(90)
set_p(120)
set_p(150)

}
