% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{VAR_RCV}
\alias{VAR_RCV}
\title{Refitted cross-validation variance estimation}
\usage{
VAR_RCV(y, x)
}
\arguments{
\item{y}{A numeric response vector of length \code{n}.}

\item{x}{A numeric design matrix with rows corresponding to observations and
columns corresponding to covariates.}
}
\value{
A list containing:
\item{sigma2}{The estimated residual variance.}
}
\description{
Estimates the residual variance in a high-dimensional linear model using
refitted cross-validation.
}
\details{
This function implements the refitted cross-validation variance estimator of
Fan et al. (2012). The data are split into two parts. A lasso model is fitted
on one part, and the selected variables are refitted on the other part to
estimate the residual variance. The procedure is then repeated in the reverse
direction, and the two variance estimates are averaged.
}
\examples{
n <- 80
p <- 100
beta <- c(sqrt(0.1 / p) * rep(1, p / 2), rep(0, p / 2))
eps <- rnorm(n)
x <- matrix(rnorm(n * p), n, p)
y <- x \%*\% beta + eps
fit <- VAR_RCV(y, x)
fit$sigma2

}
\references{
Fan, J., Guo, S., and Hao, N. (2012). Variance estimation using refitted
cross-validation in ultrahigh dimensional regression. \emph{Journal of the
Royal Statistical Society: Series B}, 74, 37--65.
}
