% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{ygztest}
\alias{ygztest}
\title{Yang, Guo, and Zhu score-based test}
\usage{
ygztest(y, u, x)
}
\arguments{
\item{y}{A numeric response vector.}

\item{u}{A numeric matrix of control covariates.}

\item{x}{A numeric matrix of covariates of interest to be tested.}
}
\value{
A list returned by \code{zctest()}, containing the standardized test
statistic and p-value.
}
\description{
Applies the score-function-based test of Yang, Guo, and Zhu (2024) by first
adjusting for control variables and then applying a high-dimensional global
test to the residuals.
}
\details{
This function is used as a benchmark method in the partial-test simulation
study. It fits a Gaussian lasso regression of \code{y} on the control covariates
\code{u} using \code{glmnet::cv.glmnet()}. The residuals are then used as the response
for testing the overall effect of the covariates of interest \code{x}.
}
\examples{
# res <- ygztest(y, u, x)

}
\references{
Yang, W., Guo, X., and Zhu, L. (2024). Score function-based tests for
ultrahigh-dimensional linear models. \emph{Electronic Journal of Statistics},
18(2), 4407--4458.
}
